requirements.txt
gitpython==3.1.43
requests==2.32.3
python-dotenv==1.0.1
openai==1.34.0
gradio==4.25.0
rich==13.7.1
